Extract contract_info queries to separate mod
