+++ crates/extrinsics/src/lib.rs
@@ -16,6 +16,7 @@
 
 mod balance;
 mod call;
+mod contract_info;
 mod env_check;
 mod error;
 mod events;
@@ -31,7 +32,6 @@ mod integration_tests;
 
 use colored::Colorize;
 use env_check::compare_node_env_with_contract;
-use subxt::utils::AccountId32;
 
 use anyhow::{
     anyhow,
@@ -40,7 +40,6 @@ use anyhow::{
 };
 use std::path::PathBuf;
 
-use crate::runtime_api::api;
 use contract_build::{
     CrateMetadata,
     DEFAULT_KEY_COL_WIDTH,
@@ -50,9 +49,11 @@ use scale::{
     Encode,
 };
 use subxt::{
+    backend::legacy::LegacyRpcMethods,
     blocks,
     config,
     tx,
+    utils::AccountId32,
     Config,
     OnlineClient,
 };
@@ -62,7 +63,6 @@ use std::{
     option::Option,
     path::Path,
 };
-use subxt::backend::legacy::LegacyRpcMethods;
 
 pub use balance::{
     BalanceVariant,
@@ -72,6 +72,12 @@ pub use call::{
     CallCommandBuilder,
     CallExec,
     CallRequest,
+};
+pub use contract_info::{
+    fetch_all_contracts,
+    fetch_contract_info,
+    fetch_wasm_code,
+    ContractInfo,
 };
 use contract_metadata::ContractMetadata;
 pub use contract_transcode::ContractMessageTranscoder;
@@ -342,36 +348,6 @@ async fn get_best_block(
         .ok_or(subxt::Error::Other("Best block not found".into()))
 }
 
-/// Fetch the contract info from the storage using the provided client.
-pub async fn fetch_contract_info(
-    contract: &AccountId32,
-    rpc: &LegacyRpcMethods<DefaultConfig>,
-    client: &Client,
-) -> Result<Option<ContractInfo>> {
-    let info_contract_call = api::storage().contracts().contract_info_of(contract);
-
-    let best_block = get_best_block(rpc).await?;
-
-    let contract_info_of = client
-        .storage()
-        .at(best_block)
-        .fetch(&info_contract_call)
-        .await?;
-
-    match contract_info_of {
-        Some(info_result) => {
-            let convert_trie_id = hex::encode(info_result.trie_id.0);
-            Ok(Some(ContractInfo {
-                trie_id: convert_trie_id,
-                code_hash: info_result.code_hash,
-                storage_items: info_result.storage_items,
-                storage_item_deposit: info_result.storage_item_deposit,
-            }))
-        }
-        None => Ok(None),
-    }
-}
-
 fn check_env_types<T>(
     client: &OnlineClient<T>,
     transcoder: &ContractMessageTranscoder,
@@ -380,103 +356,6 @@ where
     T: Config,
 {
     compare_node_env_with_contract(client.metadata().types(), transcoder.metadata())
-}
-
-#[derive(serde::Serialize)]
-pub struct ContractInfo {
-    trie_id: String,
-    code_hash: CodeHash,
-    storage_items: u32,
-    storage_item_deposit: Balance,
-}
-
-impl ContractInfo {
-    /// Convert and return contract info in JSON format.
-    pub fn to_json(&self) -> Result<String> {
-        Ok(serde_json::to_string_pretty(self)?)
-    }
-
-    /// Return the trie_id of the contract.
-    pub fn trie_id(&self) -> &str {
-        &self.trie_id
-    }
-
-    /// Return the code_hash of the contract.
-    pub fn code_hash(&self) -> &CodeHash {
-        &self.code_hash
-    }
-
-    /// Return the number of storage items of the contract.
-    pub fn storage_items(&self) -> u32 {
-        self.storage_items
-    }
-
-    /// Return the storage item deposit of the contract.
-    pub fn storage_item_deposit(&self) -> Balance {
-        self.storage_item_deposit
-    }
-}
-
-/// Fetch the contract wasm code from the storage using the provided client and code hash.
-pub async fn fetch_wasm_code(
-    client: &Client,
-    rpc: &LegacyRpcMethods<DefaultConfig>,
-    hash: &CodeHash,
-) -> Result<Option<Vec<u8>>> {
-    let pristine_code_address = api::storage().contracts().pristine_code(hash);
-    let best_block = get_best_block(rpc).await?;
-
-    let pristine_bytes = client
-        .storage()
-        .at(best_block)
-        .fetch(&pristine_code_address)
-        .await?
-        .map(|v| v.0);
-
-    Ok(pristine_bytes)
-}
-
-/// Parse a contract account address from a storage key. Returns error if a key is
-/// malformated.
-fn parse_contract_account_address(
-    storage_contract_account_key: &[u8],
-    storage_contract_root_key_len: usize,
-) -> Result<AccountId32> {
-    // storage_contract_account_key is a concatenation of contract_info_of root key and
-    // Twox64Concat(AccountId)
-    let mut account = storage_contract_account_key
-        .get(storage_contract_root_key_len + 8..)
-        .ok_or(anyhow!("Unexpected storage key size"))?;
-    AccountId32::decode(&mut account)
-        .map_err(|err| anyhow!("AccountId deserialization error: {}", err))
-}
-
-/// Fetch all contract addresses from the storage using the provided client and count of
-/// requested elements starting from an optional address
-pub async fn fetch_all_contracts(
-    client: &Client,
-    rpc: &LegacyRpcMethods<DefaultConfig>,
-) -> Result<Vec<AccountId32>> {
-    let root_key = api::storage()
-        .contracts()
-        .contract_info_of_iter()
-        .to_root_bytes();
-
-    let best_block = get_best_block(rpc).await?;
-    let mut keys = client
-        .storage()
-        .at(best_block)
-        .fetch_raw_keys(root_key.clone())
-        .await?;
-
-    let mut contract_accounts = Vec::new();
-    while let Some(result) = keys.next().await {
-        let key = result?;
-        let contract_account = parse_contract_account_address(&key, root_key.len())?;
-        contract_accounts.push(contract_account);
-    }
-
-    Ok(contract_accounts)
 }
 
 // Converts a Url into a String representation without excluding the default port.
